Start restyling the about modal
