+++ .gitignore
@@ -590,10 +590,13 @@ packages/app-mobile/components/screens/ConfigScreen/SettingsToggle.js
 packages/app-mobile/components/screens/ConfigScreen/configScreenStyles.js
 packages/app-mobile/components/screens/ConfigScreen/plugins/EnablePluginSupportPage.js
 packages/app-mobile/components/screens/ConfigScreen/plugins/PluginBox/ActionButton.js
+packages/app-mobile/components/screens/ConfigScreen/plugins/PluginBox/PluginChips.js
 packages/app-mobile/components/screens/ConfigScreen/plugins/PluginBox/PluginInfoButton.js
+packages/app-mobile/components/screens/ConfigScreen/plugins/PluginBox/PluginTitle.js
 packages/app-mobile/components/screens/ConfigScreen/plugins/PluginBox/RecommendedChip.js
 packages/app-mobile/components/screens/ConfigScreen/plugins/PluginBox/SmallChip.js
 packages/app-mobile/components/screens/ConfigScreen/plugins/PluginBox/index.js
+packages/app-mobile/components/screens/ConfigScreen/plugins/PluginInfoModal.js
 packages/app-mobile/components/screens/ConfigScreen/plugins/PluginStates.test.js
 packages/app-mobile/components/screens/ConfigScreen/plugins/PluginStates.js
 packages/app-mobile/components/screens/ConfigScreen/plugins/PluginToggle.js
